# packages/contracts/openapi/security-lab.openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FormaFlow Security Lab", "version": "1.0.0" },
  "servers": [{ "url": "http://127.0.0.1:3000" }],
  "paths": {
    "/health": { "get": { "responses": { "200": { "description": "Profil local actif" } } } },
    "/api/auth/login": { "post": { "responses": { "200": { "description": "Jeton synthétique" }, "401": { "description": "Échec maîtrisé" } } } },
    "/api/catalog": { "get": { "parameters": [{ "name": "q", "in": "query", "schema": { "type": "string", "maxLength": 120 } }], "responses": { "200": { "description": "Catalogue" } } } },
    "/api/profile": { "get": { "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Profil minimisé" } } } },
    "/api/profile/bio": { "post": { "security": [{ "bearerAuth": [] }], "responses": { "201": { "description": "Bio enregistrée" } } } },
    "/api/profile/rectification": { "post": { "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false, "required": ["displayName", "bio"], "properties": { "displayName": { "type": "string", "maxLength": 80 }, "bio": { "type": "string", "maxLength": 500 } } } } } }, "responses": { "201": { "description": "Profil propre rectifié" }, "400": { "description": "DTO refusé" } } } },
    "/api/preferences/newsletter": { "post": { "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false, "required": ["granted", "policyVersion"], "properties": { "granted": { "type": "boolean" }, "policyVersion": { "const": "privacy-notice-2026-09" } } } } } }, "responses": { "201": { "description": "Préférence facultative versionnée" }, "400": { "description": "Décision invalide" } } } },
    "/api/orders/{id}": { "get": { "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "Commande autorisée" }, "403": { "description": "Propriétaire requis" } } } },
    "/api/enrollments": { "post": { "security": [{ "bearerAuth": [] }], "responses": { "201": { "description": "Inscription et événement" } } } },
    "/api/export": { "get": { "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Export du sujet" } } } },
    "/api/account": { "delete": { "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Demande distribuée" } } } },
    "/cookie/vulnerable/login": { "post": { "responses": { "200": { "description": "Session cookie didactique faible et locale" } } } },
    "/cookie/vulnerable/account/delete": { "post": { "responses": { "201": { "description": "Démonstration CSRF locale bornée" } } } },
    "/cookie/login": { "post": { "responses": { "200": { "description": "Session cookie corrigée isolée" } } } },
    "/cookie/account/delete": { "post": { "responses": { "201": { "description": "Opération critique cookie protégée" }, "403": { "description": "Défense CSRF" } } } }
  },
  "components": { "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } } }
}